Added game manager and timer
Added a simple game manager and timer that will look for all nodes in our "People" group and add them to an array.

When these people are killed they are removed from the array, and the array length is checked to see if it is empty and thus the level was successfully beaten.

The game manager also has a public float for "Level  Time" that allows an editor to set how long the player has in that level before a game over.

ISSUES:

Currently the ui doesn't stay on the screen, it is a static entity at 0,0

There is no actual win / lose code, just a print statement when the conditions are met